move distributed simulation into subfolder
